working capturing of stdout for statusbar log

+++ tests/src/statusbarlog_test_helpers.cc
@@ -14,36 +14,53 @@
 
 // -- statusbarlog/test/src/statusbarlog_test_helpers.h
 
-#include "statusbarlog_test.h"
+// clang-format off
 
-#include <unistd.h>
 #include <fcntl.h>
 #include <sys/stat.h>
+#include <unistd.h>
 
 #include <algorithm>
-#include <filesystem>
 #include <cstring>
+#include <filesystem>
 #include <iostream>
 #include <string>
 
 #include "statusbarlog/statusbarlog.h"
-
+#include "statusbarlog_test.h"
 
 namespace statusbar_log {
-namespace test{
+namespace test {
 
-bool SetupTestOutputDirectory() {
-  std::filesystem::remove_all(test_output_dir);
-  return std::filesystem::create_directory(test_output_dir);
+void SetupTestOutputDirectory() {
+  if (kSeparateLogFiles) {
+    std::filesystem::remove_all(test_output_dir);
+    if (!std::filesystem::create_directory(test_output_dir)){
+    std::cerr << "Failed to create test output directory: " << test_output_dir
+              << std::endl;
+    }
+  }
+  if (!std::filesystem::remove_all(global_log_filename)){
+    std::cerr << "Failed to create test log file: " << global_log_filename
+              << std::endl;
+}
 }
 
 std::string GenerateTestLogFilename(const std::string& test_suite,
                                     const std::string& test_name) {
-  std::string safe_suite = test_suite;
-  std::string safe_name = test_name;
-  std::replace(safe_suite.begin(), safe_suite.end(), '/', '_');
-  std::replace(safe_name.begin(), safe_name.end(), '/', '_');
-  return test_output_dir + "/" + safe_suite + "_" + safe_name + ".log";
+  if (kSeparateLogFiles) {
+    std::string safe_suite = test_suite;
+    std::string safe_name = test_name;
+    std::replace(safe_suite.begin(), safe_suite.end(), '/', '_');
+    std::replace(safe_name.begin(), safe_name.end(), '/', '_');
+    std::replace(safe_suite.begin(), safe_suite.end(), '\\', '_');
+    std::replace(safe_name.begin(), safe_name.end(), '\\', '_');
+    return test_output_dir + "/" + safe_suite + "_" + safe_name + ".log";
+  }
+  std::string safe_file_name = global_log_filename;
+  std::replace(safe_file_name.begin(), safe_file_name.end(), '/', '_');
+  std::replace(safe_file_name.begin(), safe_file_name.end(), '\\', '_');
+  return safe_file_name;
 }
 
 int CaptureStdoutToFile(const std::string& filename) {
@@ -55,48 +72,36 @@ int CaptureStdoutToFile(const std::string& filename) {
   std::fflush(stdout);
   std::cout.flush();
 
-  // int fd = open(filename.c_str(), O_CREAT | O_WRONLY, 0644);
-  // if (fd == -1) {
-  //   std::cerr << "CaptureStdoutToFile - Error: open('" << filename
-  //             << "') failed: " << std::strerror(errno) << "\n";
-  //   return -2;
-  // }
-
+  int fd = open(filename.c_str(), O_CREAT | O_WRONLY, 0644);
+  if (fd == -1) {
+    std::cerr << "CaptureStdoutToFile - Error: open('" << filename
+              << "') failed: " << std::strerror(errno) << "\n";
+    return -2;
+  }
 
   _saved_stdout_fd = dup(STDOUT_FILENO);
   if (_saved_stdout_fd == -1) {
     std::cerr << "CaptureStdoutToFile - Error: dup(STDOUT_FILENO) failed: "
               << std::strerror(errno) << "\n";
-    // close(fd);
+    close(fd);
     return -2;
   }
 
-  FILE* f = std::freopen(filename.c_str(), "w", stdout);
-  if (f == nullptr) {
-    std::cerr << "CaptureStdoutToFile - Error: freopen('" << filename
-              << "') failed: " << std::strerror(errno) << "\n";
+  if (dup2(fd, STDOUT_FILENO) == -1) {
+    std::cerr << "CaptureStdoutToFile - Error: dup2(fd, STDOUT_FILENO) failed: "
+              << std::strerror(errno) << "\n";
+    close(fd);
     close(_saved_stdout_fd);
     _saved_stdout_fd = -1;
-    return -3;
+    return -4;
   }
-  setvbuf(stdout, nullptr, _IOFBF, 0);
 
-  // if (dup2(fd, STDOUT_FILENO) == -1) {
-  //   std::cerr << "CaptureStdoutToFile - Error: dup2(fd, STDOUT_FILENO) failed: "
-  //             << std::strerror(errno) << "\n";
-  //   close(fd);
-  //   close(_saved_stdout_fd);
-  //   _saved_stdout_fd = -1;
-  //   return -4;
-  // }
-
-  // close(fd);
+  close(fd);
   std::ios::sync_with_stdio(true);
   std::fflush(stdout);
   std::cout.flush();
 
   _is_capturing = true;
-  std::cout << "HERE 6\n" << std::endl;
   return 0;
 }
 
@@ -161,19 +166,11 @@ int RedirectDestroyStatusbarHandle(
 int RedirectUpdateStatusbar(statusbar_log::StatusbarHandle& statusbar_handle,
                             const std::size_t idx, const double percent,
                             const std::string& log_filename) {
-
-  std::cout << "starting update\n" << std::endl;
-  std::cout << "HERE 5\n" << std::endl;
   CaptureStdoutToFile(log_filename);
-  std::cout << "HERE 7\n" << std::endl;
-  std::cout << "\n\n\n";
-  std::cout.flush();
   int err_code = statusbar_log::UpdateStatusbar(statusbar_handle, idx, percent);
-  std::cout << "HERE 8\n" << std::endl;
   RestoreStdout();
-  std::cout << "HERE 9\n" << std::endl;
   return err_code;
 }
 
-} // statusbar_log
-} // test
+}  // namespace test
+}  // namespace statusbar_log
